==> notebooks/10.0-semEval16-preproc-data.ipynb <==
## SemEval16: Dataset Preprocessing 

- Author: Nana
- Created: 2022.07.21
---
- Goal: preprocess the data to fit the format provided by semEval 2022.
- Source data: SemEval 2016 Task 5 Restaurant datasets 
- XML to JSON


datadir = '../../semEval16/data'
import os, sys
queue = []
for file in os.listdir(datadir):
    if file.endswith('.xml'):
        queue.append(file)
queue

### XML structure 

from xml.etree import ElementTree as ET

filename = f'{datadir}/{queue[0]}'
tree = ET.parse(filename)
root = tree.getroot()
print(root.tag)
for review in root:
    print(review.tag, review.attrib)
    sentences = review.find('sentences')
    for sentence in sentences.findall('sentence'):
        print(sentence.attrib)
        senttext = sentence.find('text')
        opinions = sentence.find('Opinions')
        print(senttext.text)
        opinions = opinions.findall('Opinion')
        print(opinions[0].attrib)  
    break 

### SemEval 2022 Data Structure

class structs:
    def ReviewDict():
        return {
            'review_id':'', # extra, to keep the review-sentences structure in SemEval 2016 
            'sent_id': '',
            'text':'',
            'opinions':[]
        }
    def OpnDict():
        return {
        'Source':[[],[]],
        'Target':[[],[]],
        'Polar_expression':[[],[]],
        'Polarity':'',
        'Intensity':'',
        'Category': '', # extra, to keep the gold label of category in SemEval 2016 
    }

### Conversion Function

from xml.etree import ElementTree as ET
import json 

def xml2JSON(filename):
    tree = ET.parse(filename)
    root = tree.getroot()
    print(root.tag)
    REVIEWS = []
    for review in root:
        # print(review.tag, review.attrib['rid'])

        sentences = review.find('sentences')

        for sentence in sentences.findall('sentence'):
            # print('==initializing ReviewDict==')
            RD = structs.ReviewDict()
            RD['review_id'] = review.attrib['rid']
            RD['sent_id'] = sentence.attrib['id']
            RD['text'] = sentence.find('text').text

            opinions = sentence.find('Opinions')
            if opinions: 
                opinions = opinions.findall('Opinion')
                for op in opinions:
                    op = op.attrib
                    OPD = structs.OpnDict()

                    if op['target'] == 'NULL': 
                        target = [[], []]
                    else: 
                        target = [op['target'], [f"{op['from']}:{op['to']}"]]
                    # skip Source
                    OPD['Target'] = target
                    OPD['Polarity'] = op['polarity']
                    OPD['Category'] = op['category']
                    RD['opinions'].append(OPD) 
                    # print(RD['opinions'])
            else: pass 
            REVIEWS.append(RD)
            
    jsonfilename = filename.rstrip('.xml')+'.json'
    with open(jsonfilename, 'w') as fout:
        json.dump(REVIEWS , fout, ensure_ascii=False, indent = 4)
    print(f'{filename} converting to {jsonfilename}.')

for file in queue:
    filename = f'{datadir}/{file}'
    xml2JSON(filename)
