==> notebooks/20.0-s22baseline-inference-s16-scorer.ipynb <==
## Scorer: semEval22 baselines, inferences on semEval16 

    =================
    - Author: Nana 
    - Created Date: 2022.7.28
    =================
- models: two semEval22 baseline models (extracts opinion expressions `polarity_expression`
- inference dataset: semEval16 ABSA RESTAURANT dataset (with gold answer)
- metrics: micro-f1, roughly according to [SemEval-2016 Task 5: Aspect Based Sentiment Analysis](https://aclanthology.org/S16-1002/)
- tweaks: 
    - Because the s22 baselines do not return Aspect Category (e.g. FOOD#QUALITY), we cannot
    calc score completely based on the s16 metrics. Furthermore, the micro f1 relies on exact match of strings, i.e. gold: 'wine list' ; pred: 'the wine list' is counted as a false prediction,
    which might be improved using string algorithms.
    - Because the s16 gold answer does not provide `polarity_expression`, we cannot calc score based using s22 metrics either. 
    - we'll tweak the scorer function a little bit. 
    

datadir = "../semEval16/data"
seqdir = "../semEval16/sequence_labeling/predictions"
graphdir = "../semEval16/graph_parser/predictions"

splits = ['train', 'test-B', 'test-gold']

## 0.1 Check if test-B and test-gold are the same splits -> Yes

- Ans: Yes, review_id和sent_id都是同一個set。
- test-gold才是有完整答案的（全部都有標注除了intensity）
- test-B缺標polarity（缺標polarity和intensity）

splitpaths = [f'{datadir}/train/ABSA16_Restaurants_Train_SB1_v2.json',
    f'{datadir}/test-B/EN_REST_SB1_TEST-B.json',
    f'{datadir}/test-gold/EN_REST_SB1_TEST-gold.json']


import json 

splitD = {'train':None, 'test-B':None, 'test-gold':None}
for split, sppath in zip(splits, splitpaths):
    print('Loading from', sppath)
    with open(sppath, 'r') as f:
        splitD[split] = json.load(f)

Bset = set(x['sent_id'] for x in splitD['test-B'])
Gset = set(x['sent_id'] for x in splitD['test-gold']) 
Bset == Gset

len(Bset)

id = 'en_BlueRibbonSushi_478218171'
from random import randint
rindex = randint(0, len(Gset)) 
splitD['test-gold'][rindex] 

splitD['test-B'][rindex]

## 1 eval metrics
   - `Target`
       試著detect相同或相近target的opinion，將gold_target加入預測，找不到則填''。
   - `(Target, polarity)`
       試著detect相同或相近target的opinion。
       找到對應predicted polarity，將f'{gold_target}_{polarity}'加入預測，找不到則填null_null。
   - 計算f1時以tuple為單位，傳入的sklearn.metrics.f1_score(y_true, y_pred)之y_true長度應為整個dataset中gold tuples總數。
   - note that polarity is to be cast to uppercase

splits = ['train', 'test-gold']

#### loading predictions 
from collections import defaultdict
predD = defaultdict(lambda: defaultdict(dict))
# 3 dataset-pretrained models 
prt_dataset = ['darmstadt_unis', 'mpqa', 'opener_en']   
for i, baselinedir in enumerate([seqdir, graphdir]):
    baseline = 'seq' if i == 0 else 'graph'
    for split in splits:
        for pretrained_d in prt_dataset:
            fullpath = f'{baselinedir}/{split}/{pretrained_d}_prediction.json'
            with open(fullpath, 'r') as f:
                predD[baseline][split][pretrained_d] = json.load(f)

#  predD

# for the same sent_id, 
# run through gold targets 
 # run through pred targets 
    # how to ignore duplicates?
        # if lcs(gold, pred) >= 4
         # consider them to be correct and append it to the predictions 

cursplit = 'train'
baseline = 'graph'
pretrained = 'opener_en'

golden = splitD[cursplit]
prediction = predD[baseline][cursplit][pretrained]
pairing = defaultdict(lambda: {'gold':None, 'pred':None})
for y in golden:
    sid = y['sent_id']
    pairing[sid]['gold'] = y 
for y_ in prediction:
    sid = y_['sent_id']
    pairing[sid]['pred'] = y_

# credit to: https://www.geeksforgeeks.org/longest-common-substring-dp-29/
def LCSubStr(X, Y):
    m, n = len(X), len(Y)
    # Create a table to store lengths of
    # longest common suffixes of substrings.
    # Note that LCSuff[i][j] contains the
    # length of longest common suffix of
    # X[0...i-1] and Y[0...j-1]. The first
    # row and first column entries have no
    # logical meaning, they are used only
    # for simplicity of the program.
 
    # LCSuff is the table with zero
    # value initially in each cell
    LCSuff = [[0 for k in range(n+1)] for l in range(m+1)]
 
    # To store the length of
    # longest common substring
    result = 0
 
    # Following steps to build
    # LCSuff[m+1][n+1] in bottom up fashion
    for i in range(m + 1):
        for j in range(n + 1):
            if (i == 0 or j == 0):
                LCSuff[i][j] = 0
            elif (X[i-1] == Y[j-1]):
                LCSuff[i][j] = LCSuff[i-1][j-1] + 1
                result = max(result, LCSuff[i][j])
            else:
                LCSuff[i][j] = 0
    return result

def collect_targets(data):
    opns = data['opinions']
    x = set()
    for opn in opns:
        if isinstance(opn['Target'][0], str):
            x.add(opn['Target'][0])
        elif len(opn['Target'][0]) > 0:
            x.add(opn['Target'][0][0])
    return list(x)

### 1.1 target eval metrics

def target_scorer(pairing):
    '''
    pairing: defaultdict(lambda: {'gold':None, 'pred':None})
        gold_data = pairing[sent_id]['gold']
        pred_data = pairing[sent_id]['pred']
    '''
    import re
    from sklearn.metrics import f1_score

    goldtargets = [] # hold gold targets
    predtargets = [] # hold preprocessed pred targets
    correct = {'exact':0, 'lcs':0} 
    for k, pair in pairing.items():

        goldtgts = collect_targets(pair['gold'])
        predtgts = collect_targets(pair['pred'])
        # print(goldtgts, predtgts)
        for goldtarget in goldtgts:
            goldtarget = re.sub(r"[,.;@#?!&$]+\ *", "", goldtarget)
            goldtargets.append(goldtarget)

            addFlag = False
            for pi, predtarget in enumerate(predtgts):
                predtarget = re.sub(r"[,.;@#?!&$]+\ *", "", predtarget)
                # check lcs (longest common substring)
                if goldtarget == predtarget:
                    predtargets.append(goldtarget)
                    predtgts.pop(pi)
                    addFlag = True
                    correct['exact'] += 1
                    break 
                if LCSubStr(goldtarget, predtarget) >= 4:
                    # print(goldtarget, '|', predtarget)
                    predtargets.append(goldtarget)
                    predtgts.pop(pi)
                    addFlag = True
                    correct['lcs'] += 1
                    break
            if not addFlag:
                predtargets.append('')
    assert len(goldtargets) == len(predtargets)
    print('='*7+' TARGET SCORER '+'='*7)
    print(f'- # sentences: {len(pairing)}')
    print(f'- # gold tuples: {len(goldtargets)}')
    print(f'- # correct: {correct}')
    f1 = f1_score(goldtargets, predtargets, average="micro")
    print(f'- f1 score (micro): {f1}')
    return f1


target_scorer(pairing)

### 1.2 (target, pol) eval metrics

def collect_tuples(data):
    opns = data['opinions']
    x = set()
    for opn in opns:
        if isinstance(opn['Target'][0], str):
            x.add((opn['Target'][0], opn['Polarity'].lower()))
        elif len(opn['Target'][0]) > 0:
            x.add((opn['Target'][0][0], opn['Polarity'].lower()))
    return list(x)

def tuple_scorer(pairing):
    '''
    pairing: defaultdict(lambda: {'gold':None, 'pred':None})
        gold_data = pairing[sent_id]['gold']
        pred_data = pairing[sent_id]['pred']
    '''
    import re
    from sklearn.metrics import f1_score

    goldtuples = [] # hold gold targets
    predtuples = [] # hold preprocessed pred targets

    for k, pair in pairing.items():

        goldtups = collect_tuples(pair['gold'])
        predtups = collect_tuples(pair['pred'])
        for goldtarget, goldpol in goldtups:
            goldtarget = re.sub(r"[,.;@#?!&$]+\ *", "", goldtarget)
            goldtuples.append(f'{goldtarget}_{goldpol}')

            addFlag = False
            for pi, (predtarget, predpol) in enumerate(predtups):
                predtarget = re.sub(r"[,.;@#?!&$]+\ *", "", predtarget)
                # check lcs (longest common substring)
                if goldtarget == predtarget or LCSubStr(goldtarget, predtarget) >= 4:
                    predtuples.append(f'{goldtarget}_{predpol}')
                    predtups.pop(pi)
                    addFlag = True
                    break
            if not addFlag:
                predtuples.append('null_null')
    
    
    assert len(goldtuples) == len(predtuples)
    print('='*7+' TUPLE SCORER '+'='*7)
    print('- TUPLE = (target, polairty) ')
    print(f'- # sentences: {len(pairing)}')
    print(f'- # gold tuples: {len(goldtuples)}')
    # print(goldtuples[:10], predtuples[:10])
    f1 = f1_score(goldtuples, predtuples, average="micro")
    print(f'- f1 score (micro): {f1}')
    return f1


tuple_scorer(pairing)

## 2 Compute scores for each annotation 
   
   - 2 baselines x 2 pretrained datasets x 2 splits

def makepair(G, P):
    pairing = defaultdict(lambda: {'gold':None, 'pred':None})
    for y in G:
        sid = y['sent_id']
        pairing[sid]['gold'] = y 
    for y_ in P:
        sid = y_['sent_id']
        pairing[sid]['pred'] = y_
    del G, P
    return pairing             

### 2.1 score log

import numpy as np 
import pandas as pd
splits = ['train', 'test-gold']
splitstats = {}
for split in splits:
    goldfile = splitD[split]
    seqstats = pd.Series(index=prt_dataset, name = 'seq_labeling', dtype = object)
    graphstats = pd.Series(index=prt_dataset, name = 'graph_parser', dtype = object)
    for i, baselinedir in enumerate([seqdir, graphdir]):
        baseline = 'seq' if i == 0 else 'graph'
        for pretrained in prt_dataset:
            predfile = predD[baseline][split][pretrained]
            # making pairing dictionary 
            print(f'\nNOW EVALUATING {baseline}_{split}_{pretrained}...\n')
            pairing = makepair(goldfile, predfile)
            tgtf1 = round(target_scorer(pairing), 4)
            tupf1 = round(tuple_scorer(pairing), 4) 
            if i == 0:
                seqstats[pretrained] = (tgtf1, tupf1)
            else: graphstats[pretrained] = (tgtf1, tupf1)
    splitstats[split] = pd.concat([seqstats, graphstats], axis=1)
    

### 2.2 score tables
score description: `(target micro f1, target_polarity tuple micro f1)`

splitstats['train']

splitstats['test-gold']

s
