==> notebooks/23.0-json-to-csv.ipynb <==
import json

with open('../semEval16/sequence_labeling/predictions/train/opener_en_prediction.json') as f:
    data = json.load(f)

data[0:3]

len(data)

alldata = []

for i in range(len(data)):
    
    if len(data[i]['opinions']) == 0:
        # print(data[i]['sent_id'])
        seq = ('sent_id', 'text', 'Source', 'Target', 'Polar_expression', 'Polarity')
        d = dict.fromkeys(seq)
        
        d['sent_id'] = data[i]['sent_id']
        d['text'] = data[i]['text']
        alldata.append(d)
        
    else:       
        for j in range(len(data[i]['opinions'])):
            # print(data[i]['sent_id'], data[i]['opinions'][j])
            seq = ('sent_id', 'text', 'Source', 'Target', 'Polar_expression', 'Polarity')
            d = dict.fromkeys(seq)

            d['sent_id'] = data[i]['sent_id']
            d['text'] = data[i]['text']
        
            keys = ['Source', 'Target', 'Polar_expression']
            
            for k in keys:
                if len(data[i]['opinions'][j][k][0]) != 0:
                    d[k] = data[i]['opinions'][j][k][0][0]
                else:
                    d[k] = Noneasdzxc8914
                    
                
            d['Polarity'] = data[i]['opinions'][j]['Polarity']
            alldata.append(d)

alldata[10:20]

import pandas as pd

df = pd.DataFrame.from_records(alldata)
df

df.to_csv('../semEval16/check_correctness/seq_train.csv', index=False)
