==> notebooks/30.0-target-auto-matching.ipynb <==
import nltk
nltk.download('stopwords')
from nltk.corpus import stopwords
import pandas as pd
import json

# example of predicted and golden target
pred_tar = "The combination"
gold_tar = "The combination of super-fresh ingredients"

pred_tokens = nltk.word_tokenize(pred_tar.lower())
gold_tokens = nltk.word_tokenize(gold_tar.lower())

print(pred_tokens)
print(gold_tokens)

# remove stopwords
stopword = stopwords.words('english')  

pred_ls = [k for k in pred_tokens if k not in stopword]
gold_ls = [k for k in gold_tokens if k not in stopword]

print(pred_ls)
print(gold_ls)

df_pred = pd.read_csv('../data/semEval16/check_correctness/graph_test.csv')
df_pred['Target_matching'] = False
df_pred

# open json file of gold answer 
with open('../data/semEval16/raw_data/test-gold/EN_REST_SB1_TEST-gold.json') as f:
    gold = json.load(f)

gold[0:10]

for i in gold:
    for j in i['opinions']:
        if len(j['Target'][0]) > 0:
            print(j['Target'][0])

##### Matching Rule: at least one common word between predicted target and golden target

matching_result = []

for p in range(len(df_pred)):
    id = df_pred.iloc[p]['sent_id']
    matching_result.append(False)
    
    if pd.isnull(df_pred.iloc[p]['Target']) == False:
        
        for g in gold:
            if id == g['sent_id']:

                for o in g['opinions']:
                    if len(o['Target'][0]) > 0:
                        pstr = df_pred.iloc[p]['Target']
                        gstr = o['Target'][0]

                        ptokens = nltk.word_tokenize(pstr.lower())
                        gtokens = nltk.word_tokenize(gstr.lower())

                        plst = [k for k in ptokens if k not in stopword]
                        glst = [m for m in gtokens if m not in stopword]
                        # print(plst, glst)

                        for pelement in plst:
                            if pelement in glst:
                                matching_result[p] = True
                                
            
# for g in gold:
#     for o in g['opinions']:
#         print(o['Target'][0], len(o['Target'][0]))

# for p in range(20):
#     if (pd.isnull(df_pred.iloc[p]['Target'])) == False:
#         print(df_pred.iloc[p]['Target'])

df_pred['Target_matching'] = matching_result
df_pred[df_pred.Target_matching == True]

df_pred.to_csv("../data/semEval16/check_correctness/graph-test-correctness-auto.csv", index=False)
