==> notebooks/31.5-yelp-2-s22format.ipynb <==
## YELP data to S22 format
- Author: Nana
- Created: 2022.08.22
---
- Goal: preprocess yelp to fit the format provided by semEval 2022.
- Source data: `/share/data/yelp/yelp_academic_dataset_review.json`
- Converted data: `/share/data/nana2929/yelp/`


from utils import structs
import json 
import numpy as np

# wow are we turning everything to ONE File? 
# splitting to 3 files? 

YELPPATH = '/share/data/yelp'
RP = 'yelp_academic_dataset_review.json'
rppath = f'{YELPPATH}/{RP}'
yelpdir = '/share/home/nana2929/yelp'

REVIEWS = []
with open (rppath, 'r') as f:
    for id, line in enumerate(f):
        line = json.loads(line)
        if id == 0: print(line)
        REVIEWS.append(line) 

def transform(x):
    if x == 3: return 'Neutral'
    elif x < 3: return 'Negative'
    else: return 'Positive'

def Yelp2JSON(reviews):
    '''
    input: YELP a list of dictionaries
    return: SemEval22 format json dictionaries (with positivity)
    '''
    result = []
    for review in reviews:
        RD = structs.ReviewDict()
        RD['review_id'] = review['review_id']
        RD['text'] = review['text'].replace('\n', '')
        RD['overall_polarity'] = transform(int(review['stars']))
        result.append(RD)
    return result 

### Because of the large size of YELP, split the files to avoid OOM/crashes

NSPLIT = 6

N_REVIEWS = np.array_split(REVIEWS, NSPLIT)

import time
for i, split in enumerate(N_REVIEWS):
    converted_split = Yelp2JSON(split)
    jsonfilename = f'{yelpdir}/s22_formatted_yelp_reviews_{i+1}.json'
    with open(jsonfilename, 'w') as fout:
        json.dump(converted_split, fout, ensure_ascii = False, indent = 4)
    print(f'{len(split)} data is successfully converted to {jsonfilename}.')

# display file size

%%bash
cd ../../yelp
ls -lh *

# try reading in 
with open(f'{yelpdir}/s22_formatted_yelp_reviews_1.json', 'r') as fin:
    x = json.load(fin)
x[0]
