==> notebooks/32.0-semeval16-csv-convert.ipynb <==
import pandas as pd
import json

from fuzzywuzzy import fuzz
from fuzzywuzzy import process

# similariy test
fuzz.ratio("Where can I begin.", \
    "– Where can I begin.")


##### Mistakes of correctness.csv
###### 1. Some pred_target's format and length are incorrect due to wrong conversion
###### 2. 'sent_id' of 'graph-test-correctness.csv' are all wrong due to wrong sorting process
###### 3. It's necessary to adjust prediction results (json files) manually in certain steps 

df = pd.read_csv('../data/semEval16/check_correctness/graph-test-correctness.csv', encoding='utf-8')
df[df.Target_matching == True]

# open json file of training data or test data
with open('../data/semEval16/graph_parser/predictions/test-gold/opener_en_prediction.json') as f:
    json_data = json.load(f)

json_data[:10]

# run this cell only for test data
# adjust sent_id from test-correctness

c = 0
for i in range(len(df)):
    flag = False

    for j in json_data:
        if df.iloc[i]['text'] == j['text'] or fuzz.ratio(df.iloc[i]['text'], j['text']) >= 95:
            c += 1
            flag = True
            df.iloc[i,0] = j['sent_id'] # edit sent_id
            break
            
    if flag == False:
        print(i, df.iloc[i]['text'])

print(f'number of df rows: {len(df)}')
print(f'number of df rows with correct sent_id: {c}')

# assign 'Matching' to json based on correctness.csv(df)

count = 0

for i in json_data:
    sid = i['sent_id'] 

    for j in i['opinions']:
        j['Matching'] = False

        if len(j['Target'][0]) > 0:
            tar = j['Target'][0][0]
            exp = j['Polar_expression'][0][0]
            df2 = df[df.sent_id == sid]

            for k in range(len(df2)):
                # match pairs of json and csv
                if (pd.isnull(df2.iloc[k]['Target']) == False) and (df2.iloc[k]['Target'] == tar) and (df2.iloc[k]['Polar_expression'] == exp):
                    # print(df2.iloc[k]['Target'], df2.iloc[k]['Polar_expression'])   
                    if (df2.iloc[k]['Target_matching'] == True):
                        j['Matching'] = True
                        count += 1
                        break
                            
        
print(f'number of true targets: {count}')

# run this cell only for test data
# same pair but different sent_id

for i in range(len(df)):
    if (df.iloc[i]['text'] == 'Great Breakfast'):
        print(i)
        print(df.iloc[i])

# run this cell only for test data
# edit duplicated sent_id

df.iloc[288,0] = "en_SnoozeanAMEatery_479991844:0"

# deep copy original json file
import copy
json_list = copy.deepcopy(json_data)

# check true pairs and print missing true pairs in json

true_df = df[df.Target_matching == True]
wrong = 0
sid_true = []

for s in range(len(true_df)):
    flag = False

    for t in range(len(json_list)):
        for r in json_list[t]['opinions']:
            if (r['Matching']) and (true_df.iloc[s]['sent_id'] == json_list[t]['sent_id']):
                if  (true_df.iloc[s]['Target'] == r['Target'][0][0]) and (true_df.iloc[s]['Polar_expression'] == r['Polar_expression'][0][0]):
                    flag = True
                    break
        if flag == True:
            break


    if flag == False:
        wrong += 1
        sid_true.append(true_df.iloc[s]['sent_id'])
        print(true_df.iloc[s]['sent_id'] + ', ' + true_df.iloc[s]['Target'] + ', ' + true_df.iloc[s]['Polar_expression'])

print(f'number of missing pairs: {wrong}')

# correct 'Matching' results manually
for i in range(len(json_list)):
    if json_list[i]['sent_id'] in sid_true:
        for j in range(len(json_list[i]['opinions'])):
            print(i,j,json_list[i]['opinions'][j])

json_list[497]['opinions'][0]['Matching'] = True
json_list[557]['opinions'][1]['Matching'] = True
json_list[559]['opinions'][0]['Matching'] = True
json_list[559]['opinions'][1]['Matching'] = True
json_list[563]['opinions'][0]['Matching'] = True

# check numbe of true pairs

count = 0
for i in range(len(json_list)):
        for j in json_list[i]['opinions']:
            if j['Matching']:
                count += 1

print(f'number of true targets: {count}')

# remove all false pairs

for i in range(len(json_list)):
    new_list = [j for j in json_list[i]['opinions'] if j['Matching']]
    json_list[i]['opinions'] = new_list
        
json_list[:10]

# remove all empty opinions

json_new = []

for i in range(len(json_list)):
    if len(json_list[i]['opinions']) > 0:
        json_new.append(json_list[i])

print(f'number of text: {len(json_new)}')

# add key of dataset

for i in range(len(json_new)):
    json_new[i]['dataset'] = 's16+_test'

json_new[:10]

# convert to json file

with open('../data/semEval16_plus/s16+_test.json', 'w') as outfile:
    json.dump(json_new, outfile, indent = 4, ensure_ascii=False)
