==> notebooks/37.0-regulate-polar-major.ipynb <==
import pandas as pd
import json

with open('../data/yelp/ANS_polar-major_2rest100.json') as f:
    gold1 = json.load(f)

gold1[:2]

# count number of opinion
opnum = 0

for g in range(len(gold1)):
    if len(gold1[g]['opinions']) == 1:
        opnum += len(gold1[g]['opinions'])

opnum


# move 'polar_ans' into 'opinions'
# set 'polar_ans' = 'default' if length of opinions > 1
for g in range(len(gold1)):
    if (len(gold1[g]['opinions']) >= 1):  
        for op in range(len(gold1[g]['opinions'])):
            if len(gold1[g]['opinions'][op]['Target'][0]) == 0: # no 'Target' -> 'polar_ans' = false
                gold1[g]['opinions'][op]['Polar_ans'] = False
            else:
                if (len(gold1[g]['opinions']) == 1):
                    gold1[g]['opinions'][op]['Polar_ans'] = gold1[g]['polar_ans']
                else:
                    gold1[g]['opinions'][op]['Polar_ans'] = "default"
                # print(gold1[g]['opinions'][op])

with open('../data/yelp/RAW_yelp-food-test_2rest100.json') as f:
    raw = json.load(f)

raw[:2]

for i in raw[:5]:
    print(i['review_id'], i['user_id'])

# add two key: 'user_id' & 'date'
# delete 'polar_ans'
for g in range(len(gold1)):
    for i in raw:
        if i['review_id'] == gold1[g]['review_id']:
            gold1[g]['user_id'] = i['user_id']
            gold1[g]['date'] = i['date']
            break
    del gold1[g]['polar_ans']
    # print(gold1[g]['review_id'])

gold1[:3]

# output json 
with open('../data/yelp/ANS2-raw_polar-major_2rest100.json', 'w') as outfile:
    json.dump(gold1, outfile, indent = 4, ensure_ascii=False)

n = 0
for g in range(len(gold1)):
    if (len(gold1[g]['opinions']) >= 2): 
        n += 1

print(f'{n} text with length of opinions > 1')

### After regulating polar_ans, convert the json to csv file

with open('../data/yelp/ANS2-correct_polar-major_2rest100.json') as f:
    gold2 = json.load(f)

gold2[:2]

# df column: rest | user_id | review_id | sent_id | text | Target | Target_key | Polar_expression | Polarity | Polar_ans | date | stars
# rest1's bus_id: '_ab50qdWOk0DdB6XOrBitw'

rest, user_id, review_id, sent_id, text= [], [], [], [], []
Target, Target_key, Polar_expression, Polarity, Polar_ans = [], [], [], [], []
date, stars = [], []

for i in gold2:
    if len(i['opinions']) > 0:
        for j in i['opinions']:
            if i['business_id'] == '_ab50qdWOk0DdB6XOrBitw':
                rest.append(1)
            else:
                rest.append(2)

            user_id.append(i['user_id'])
            review_id.append(i['review_id'])
            sent_id.append(i['sent_id'])
            text.append(i['text'])

            Target.append(j['Target'][0])
            Target_key.append(j['Target_key'])
            Polar_expression.append(j['Polar_expression'][0])
            Polarity.append(j['Polarity'])
            Polar_ans.append(j['Polar_ans'])

            date.append(i['date'])
            stars.append(i['stars'])
            # print(j['Target'][0], j['Target_key'], j['Polar_expression'][0])

df = pd.DataFrame({'rest': rest, 'user_id': user_id, 'review_id': review_id, 'sent_id': sent_id, 'text': text,
                    'Target': Target, 'Target_key': Target_key, 'Polar_expression': Polar_expression, 'Polarity': Polarity,
                    'Polar_ans': Polar_ans, 'date': date, 'stars': stars})
df

len(df[(df.Polar_ans == False) &  (df.Polarity == 'Positive')] )

# df to csv
df.to_csv('../data/yelp/STAT2_polar-table_2rest.csv', index=False)
